Work with `sudo -i`.  Closes #61.
The fist part of this fix is to honor the HOME environment variable, which sudo(8) manages appropriately.  This removes reliance on SUDO_USER for determining the correct home directory.

The second part of this fix is to inspect SUDO_COMMAND to determine indirectly whether the command was run under `sudo -i`.  If SUDO_COMMAND mentions "blueprint" then it is assumed to be a one-off and the original uid/gid are used for filesystem operations.  Otherwise, 0/0 are used.