Added -r to ignore version numbers.  Closes #36.
The easy part of this patch is propagating the relaxed flag through all the code generation routines, driven by the -r and --relaxed arguments to blueprint-apply, blueprint-create, and blueprint-show.

The hard part is in the shell code generator, where the manager needed to be decomposed into a generator for the gating code and another for the installation code so SERVICE_*_* variables could still be set.