Construct Blueprint objects by walking.
The blueprint JSON format is reconstituted in Python as a dict-of-dicts-and-lists but in reality the format involves a lot of defaultdict and set objects, which have nicer APIs for use in live Blueprint objects.  The constructor now uses the walk algorithm with a set of callbacks that reconstitute the rich object structure with the same data.